+++ src/selection/index.ts
@@ -1,4 +1,4 @@
-export { Lasso } from './Lasso';
-export type { LassoProps, LassoType } from './LassoContent';
+export { LassoWrapper as Lasso } from './LassoWrapper';
+export type { LassoProps, LassoType } from './Lasso';
 export * from './useSelection';
 export * from './utils';
